#2884: refactor snapshot business out of core -> factus
