+++ factcast-factus/src/main/java/org/factcast/factus/FactusImpl.java
@@ -39,12 +39,10 @@ import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.checkerframework.checker.units.qual.A;
 import org.factcast.core.Fact;
 import org.factcast.core.FactCast;
 import org.factcast.core.FactStreamPosition;
 import org.factcast.core.event.EventConverter;
-import org.factcast.core.snap.Snapshot;
 import org.factcast.core.spec.FactSpec;
 import org.factcast.core.subscription.Subscription;
 import org.factcast.core.subscription.SubscriptionRequest;
@@ -63,6 +61,7 @@ import org.factcast.factus.projector.ProjectorFactory;
 import org.factcast.factus.serializer.SnapshotSerializer;
 import org.factcast.factus.snapshot.AggregateSnapshotRepository;
 import org.factcast.factus.snapshot.ProjectionSnapshotRepository;
+import org.factcast.factus.snapshot.Snapshot;
 import org.factcast.factus.snapshot.SnapshotSerializerSupplier;
 
 /** Single entry point to the factus API. */
