+++ factcast-factus/src/main/java/org/factcast/factus/snapshot/ProjectionSnapshotRepositoryImpl.java
@@ -19,9 +19,6 @@ import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 import lombok.NonNull;
-import org.factcast.core.snap.Snapshot;
-import org.factcast.core.snap.SnapshotCache;
-import org.factcast.core.snap.SnapshotId;
 import org.factcast.factus.metrics.FactusMetrics;
 import org.factcast.factus.projection.SnapshotProjection;
 import org.factcast.factus.serializer.SnapshotSerializer;
