+++ factcast-factus/src/main/java/org/factcast/factus/snapshot/SnapshotCache.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.factcast.core.snap;
+package org.factcast.factus.snapshot;
 
 import java.util.Optional;
 import lombok.NonNull;
