+++ factcast-factus/src/main/java/org/factcast/factus/snapshot/SnapshotId.java
@@ -13,20 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.factcast.core.snap;
+package org.factcast.factus.snapshot;
 
 import java.io.Serializable;
 import java.util.UUID;
 import lombok.NonNull;
 import lombok.Value;
 
-@Value
-public class Snapshot implements Serializable {
-  @NonNull SnapshotId id;
+@Value(staticConstructor = "of")
+public class SnapshotId implements Serializable {
+  private static final long serialVersionUID = -3207528229703207635L;
+  @NonNull String key;
 
-  @NonNull UUID lastFact;
-
-  @NonNull byte[] bytes;
-
-  boolean compressed;
+  @NonNull UUID uuid; // semantically optional, might be constant if not needed
+  // (SnapshotProjections)
 }
