+++ factcast-factus/src/test/java/org/factcast/factus/FactusImplTest.java
@@ -41,8 +41,6 @@ import org.factcast.core.FactCast;
 import org.factcast.core.FactStreamPosition;
 import org.factcast.core.TestFact;
 import org.factcast.core.event.EventConverter;
-import org.factcast.core.snap.Snapshot;
-import org.factcast.core.snap.SnapshotId;
 import org.factcast.core.spec.FactSpec;
 import org.factcast.core.subscription.Subscription;
 import org.factcast.core.subscription.SubscriptionRequest;
@@ -64,6 +62,8 @@ import org.factcast.factus.projector.ProjectorImpl;
 import org.factcast.factus.serializer.SnapshotSerializer;
 import org.factcast.factus.snapshot.AggregateSnapshotRepository;
 import org.factcast.factus.snapshot.ProjectionSnapshotRepository;
+import org.factcast.factus.snapshot.Snapshot;
+import org.factcast.factus.snapshot.SnapshotId;
 import org.factcast.factus.snapshot.SnapshotSerializerSupplier;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
