+++ factcast-factus/src/test/java/org/factcast/factus/snapshot/AggregateSnapshotRepositoryTest.java
@@ -25,7 +25,6 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.atomic.AtomicReference;
 import lombok.NonNull;
-import org.factcast.core.snap.Snapshot;
 import org.factcast.factus.projection.Aggregate;
 import org.junit.jupiter.api.Test;
