+++ factcast-factus/src/test/java/org/factcast/factus/snapshot/ProjectionSnapshotRepositoryImplTest.java
@@ -24,9 +24,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
-import org.factcast.core.snap.Snapshot;
-import org.factcast.core.snap.SnapshotCache;
-import org.factcast.core.snap.SnapshotId;
 import org.factcast.factus.metrics.FactusMetrics;
 import org.factcast.factus.projection.SnapshotProjection;
 import org.factcast.factus.serializer.ProjectionMetaData;
