+++ factcast-itests/factcast-itests-factus/src/test/java/org/factcast/itests/factus/client/RedissonSnapshotCacheTest.java
@@ -15,7 +15,7 @@
  */
 package org.factcast.itests.factus.client;
 
-import org.factcast.core.snap.SnapshotCache;
+import org.factcast.factus.snapshot.SnapshotCache;
 import org.factcast.itests.TestFactusApplication;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
