+++ factcast-itests/factcast-itests-factus/src/test/java/org/factcast/itests/factus/client/SnapshotCacheTest.java
@@ -21,9 +21,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.factcast.core.snap.Snapshot;
-import org.factcast.core.snap.SnapshotCache;
-import org.factcast.core.snap.SnapshotId;
+import org.factcast.factus.snapshot.Snapshot;
+import org.factcast.factus.snapshot.SnapshotCache;
+import org.factcast.factus.snapshot.SnapshotId;
 import org.factcast.test.AbstractFactCastIntegrationTest;
 import org.junit.jupiter.api.Test;
