--- factcast-snapshotcache-local-disk/src/main/java/org/factcast/core/snap/local/InMemoryAndDiskSnapshotProperties.java
@@ -1,33 +1,0 @@
-/*
- * Copyright © 2017-2023 factcast.org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.factcast.core.snap.local;
-
-import lombok.Data;
-import lombok.experimental.Accessors;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@Data
-@Slf4j
-@Accessors(fluent = false)
-@ConfigurationProperties(prefix = InMemoryAndDiskSnapshotProperties.PROPERTIES_PREFIX)
-public class InMemoryAndDiskSnapshotProperties {
-  public static final String PROPERTIES_PREFIX = "factcast.local.disk";
-
-  private int deleteSnapshotStaleForDays = 90;
-}
