+++ factcast-snapshotcache-local-memory/pom.xml
@@ -39,7 +39,7 @@
     </dependency>
     <dependency>
       <groupId>org.factcast</groupId>
-      <artifactId>factcast-core</artifactId>
+      <artifactId>factcast-factus</artifactId>
     </dependency>
 
     <dependency>
