+++ factcast-snapshotcache-local-memory/src/main/java/org/factcast/core/snap/local/InMemorySnapshotCache.java
@@ -20,9 +20,9 @@ import com.google.common.cache.CacheBuilder;
 import java.time.Duration;
 import java.util.Optional;
 import lombok.NonNull;
-import org.factcast.core.snap.Snapshot;
-import org.factcast.core.snap.SnapshotCache;
-import org.factcast.core.snap.SnapshotId;
+import org.factcast.factus.snapshot.Snapshot;
+import org.factcast.factus.snapshot.SnapshotCache;
+import org.factcast.factus.snapshot.SnapshotId;
 
 public class InMemorySnapshotCache implements SnapshotCache {
