+++ factcast-snapshotcache-local-memory/src/test/java/org/factcast/core/snap/local/InMemorySnapshotCacheTest.java
@@ -18,8 +18,8 @@ package org.factcast.core.snap.local;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.UUID;
-import org.factcast.core.snap.Snapshot;
-import org.factcast.core.snap.SnapshotId;
+import org.factcast.factus.snapshot.Snapshot;
+import org.factcast.factus.snapshot.SnapshotId;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
