+++ factcast-snapshotcache-redisson/src/main/java/org/factcast/core/snap/redisson/RedissonSnapshotCache.java
@@ -21,9 +21,9 @@ import java.util.*;
 import java.util.concurrent.*;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
-import org.factcast.core.snap.Snapshot;
-import org.factcast.core.snap.SnapshotCache;
-import org.factcast.core.snap.SnapshotId;
+import org.factcast.factus.snapshot.Snapshot;
+import org.factcast.factus.snapshot.SnapshotCache;
+import org.factcast.factus.snapshot.SnapshotId;
 import org.redisson.api.RBucket;
 import org.redisson.api.RMap;
 import org.redisson.api.RedissonClient;
