+++ factcast-spring-boot-autoconfigure/src/main/java/org/factcast/spring/boot/autoconfigure/factus/FactusAutoConfiguration.java
@@ -20,7 +20,6 @@ import java.util.*;
 import lombok.extern.slf4j.Slf4j;
 import org.factcast.core.FactCast;
 import org.factcast.core.event.EventConverter;
-import org.factcast.core.snap.SnapshotCache;
 import org.factcast.factus.Factus;
 import org.factcast.factus.FactusImpl;
 import org.factcast.factus.event.EventSerializer;
@@ -33,6 +32,7 @@ import org.factcast.factus.serializer.DefaultSnapshotSerializer;
 import org.factcast.factus.serializer.SnapshotSerializer;
 import org.factcast.factus.snapshot.AggregateSnapshotRepositoryImpl;
 import org.factcast.factus.snapshot.ProjectionSnapshotRepositoryImpl;
+import org.factcast.factus.snapshot.SnapshotCache;
 import org.factcast.factus.snapshot.SnapshotSerializerSupplier;
 import org.factcast.factus.utils.FactusDependency;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
