+++ factcast-spring-boot-autoconfigure/src/main/java/org/factcast/spring/boot/autoconfigure/snap/InMemorySnapshotCacheAutoConfiguration.java
@@ -15,10 +15,10 @@
  */
 package org.factcast.spring.boot.autoconfigure.snap;
 
-import org.factcast.core.snap.SnapshotCache;
 import org.factcast.core.snap.local.InMemorySnapshotCache;
 import org.factcast.core.snap.local.InMemorySnapshotProperties;
 import org.factcast.factus.Factus;
+import org.factcast.factus.snapshot.SnapshotCache;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureOrder;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
